improve stats charts
